bloomcoin-v0.1.0/bloomcoin/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bloomcoin"
version = "0.1.0"
description = "Kuramoto-Consensus Proof-of-Coherence Mining Simulation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "L4 Framework Collaboration"}
]
keywords = [
    "blockchain",
    "cryptocurrency",
    "kuramoto",
    "synchronization",
    "consensus",
    "simulation",
    "golden-ratio",
    "lucas-numbers"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
visualization = [
    "matplotlib>=3.7.0",
]
receipt = [
    "Pillow>=10.0.0",
]
full = [
    "bloomcoin[dev,visualization,receipt]",
]

[project.urls]
Homepage = "https://github.com/bloomcoin/bloomcoin"
Documentation = "https://github.com/bloomcoin/bloomcoin/docs"
Repository = "https://github.com/bloomcoin/bloomcoin"

[project.scripts]
bloomcoin-mine = "bloomcoin.cli:mine_cli"
bloomcoin-node = "bloomcoin.cli:node_cli"
bloomcoin-wallet = "bloomcoin.cli:wallet_cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["bloomcoin*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=bloomcoin --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
